# docs/swagger.yaml
definitions:
...
  api.createUserRequest:
    properties:
      address:
        type: string
      driving_license_number:
        type: string
      email:
        type: string
      first_name:
        type: string
      goverment_id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
...
  api.deleteUserResponse:
    type: object
  api.getUserResponse:
    properties:
      ID:
        type: integer
      address:
        type: string
      driving_license_number:
        type: string
      email:
        type: string
      first_name:
        type: string
      goverment_id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
  api.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  api.loginResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/api.userResponse'
    type: object
...
  api.userResponse:
    properties:
      ID:
        type: integer
      address:
        type: string
      driving_license_number:
        type: string
      email:
        type: string
      first_name:
        type: string
      goverment_id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This API serves as a backend for Vehicle Management System
  title: Vehicle Management System API
  version: "1.0"
paths:
...
  /login:
    post:
      description: Logs user in
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginResponse'
      summary: Login user
      tags:
      - user
...
  /user:
    get:
      description: Gets all users from database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.getUserResponse'
      security:
      - ApiKeyAuth: []
      summary: Get users
      tags:
      - user
    post:
      description: Creates and saves user to database
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
      security:
      - ApiKeyAuth: []
      summary: Create user
      tags:
      - user
  /user/{id}:
    delete:
      description: Deletes user from database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.deleteUserResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - user
    get:
      description: Gets user from database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.getUserResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user
      tags:
      - user
    put:
      description: Updates user in database
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - user
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"